test: find context info req, resp, session using extractor class
